requirements.txt
google-genai>=1.10.0
python-dotenv==1.0.1
flask==3.0.2